2.x: scalar flatMap optimization enabled
+ I wanted to make sure the slow path passes all tests before enabling
the scalar optimization.
+ Fixed headers for internal classes
