+++ src/main/java/io/reactivex/internal/operators/OperatorFlatMap.java
@@ -1,17 +1,14 @@
-/*
- * Copyright 2011-2015 David Karnok
+/**
+ * Copyright 2015 Netflix, Inc.
  * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in
+ * compliance with the License. You may obtain a copy of the License at
  * 
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is
+ * distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See
+ * the License for the specific language governing permissions and limitations under the License.
  */
 
 package io.reactivex.internal.operators;
@@ -134,8 +131,8 @@ public final class OperatorFlatMap<T, U> implements Operator<U, T> {
                 onError(e);
                 return;
             }
-            if (p instanceof PublisherScalarSource) {
-                tryEmitScalar(((PublisherScalarSource<? extends U>)p).value());
+            if (p instanceof ObservableScalarSource) {
+                tryEmitScalar(((ObservableScalarSource<? extends U>)p).value());
             } else {
                 InnerSubscriber<T, U> inner = new InnerSubscriber<>(this, uniqueId++);
                 addInner(inner);
@@ -220,7 +217,11 @@ public final class OperatorFlatMap<T, U> implements Operator<U, T> {
                         s.request(1);
                     }
                 } else {
-                    
+                    Queue<U> q = getMainQueue();
+                    if (!q.offer(value)) {
+                        onError(new IllegalStateException("Scalar queue full?!"));
+                        return;
+                    }
                 }
                 if (decrementAndGet() == 0) {
                     return;
